feat: Batting 전략 인터페이스 추가
