+++ src/main/java/com/hyunec/cosmicbaseballinit/batting/BattingController.java
@@ -1,10 +1,10 @@
-package com.hyunec.cosmicbaseballinit.batter;
+package com.hyunec.cosmicbaseballinit.batting;
 
-import lombok.AllArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@AllArgsConstructor
+@RequiredArgsConstructor
 @RestController
 public class BattingController {
