+++ src/main/java/com/hyunec/cosmicbaseballinit/batting/BattingResultService.java
@@ -1,11 +1,9 @@
-package com.hyunec.cosmicbaseballinit.batter;
+package com.hyunec.cosmicbaseballinit.batting;
 
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 @Slf4j
-@RequiredArgsConstructor
 @Service
 public class BattingResultService {
