+++ src/main/java/com/hyunec/cosmicbaseballinit/batting/BattingStrategy.java
@@ -1,0 +1,7 @@
+package com.hyunec.cosmicbaseballinit.batting;
+
+public interface BattingStrategy {
+
+    BattingResults generateResult();
+
+}
